Implemented channel boundary conditions
ghost floes are created for any floe that crosses the boundary x=+-Lx; their forces/torques are computed and added to their parent floes (only interaction points that exit in the nterior of the domain are kept to avoid double-counting of forces). When the floe coordinate goes out of bounds it shifts to another side of the domain.